# tools/camofox-browser/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "camofox-browser",
    "version": "1.8.15",
    "description": "Anti-detection browser automation server for AI agents. Accessibility snapshots, element refs, session isolation, cookie import, proxy rotation, and structured logs.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Jo Inc",
      "url": "https://askjo.ai",
      "email": "oss@askjo.ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9377",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token matching CAMOFOX_API_KEY (per-route auth for sensitive endpoints like cookie import and traces)."
      },
      "AccessKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token matching CAMOFOX_ACCESS_KEY. When set, gates all routes except /health, cookie import, and /stop. Acts as a superkey -- also accepted by endpoints that normally require CAMOFOX_API_KEY."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/tabs/{tabId}/wait": {
      "post": {
        "tags": [
          "Interaction"
        ],
        "summary": "Wait for a selector or timeout",
        "parameters": [
          {
            "name": "tabId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "selector": {
                    "type": "string"
                  },
                  "timeout": {
                    "type": "integer",
                    "description": "Max wait in ms."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wait completed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tab not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tabs/{tabId}/press": {
      "post": {
        "tags": [
          "Interaction"
        ],
        "summary": "Press a keyboard key",
        "parameters": [
          {
            "name": "tabId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "key"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string",
                    "description": "Key name (e.g. \"Enter\", \"Escape\", \"Tab\")."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Key pressed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tab not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tabs/{tabId}/scroll": {
      "post": {
        "tags": [
          "Interaction"
        ],
        "summary": "Scroll the page",
        "parameters": [
          {
            "name": "tabId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "direction": {
                    "type": "string",
                    "description": "\"up\" or \"down\" (default \"down\")."
                  },
                  "amount": {
                    "type": "integer",
                    "description": "Pixels to scroll."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scroll result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tab not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}